Update authors report with new flags
The `libro report --author` produces a list of books read by author
grouped by the count of books. It was hard coded to limit to 3 books
by author to show and a read_date must of been set to include.

These change introduces to flags to modify that behavior

Add --limit flag to specify limit how many books to show for author, the
default is still 3 if not specified.

Add --undated flag to specify not required read_date to be set, will use
all books in DB regardless of read_date or not